Refactored and unified 2D and 3D matrices
