+++ deep_conmech/simulator/setting/setting_matrices.py
@@ -5,95 +5,13 @@ import numba
 import numpy as np
 from conmech.features.boundaries import Boundaries
 from conmech.helpers import nph
+from deep_conmech.simulator.matrices.matrices_2d import get_edges_features_matrix_2d_numba, get_matrices
 from deep_conmech.simulator.setting.setting_mesh import SettingMesh
 from numba import njit
 
 
-@njit  # (parallel=True)
-def get_edges_features_matrix_2d_numba(elements, nodes):
-    ELEMENT_NODES_COUNT = 3
-    CONNECTED_EDGES_COUNT = 2
-    int_Ph = 1 / ELEMENT_NODES_COUNT
-    # integral of phi over the element (in 2D: 1/3, in 3D: 1/4)
-    nodes_count = len(nodes)
-    elements_count, element_size = elements.shape
-
-    edges_features_matrix = np.zeros((8, nodes_count, nodes_count), dtype=np.double)
-    element_initial_volume = np.zeros(elements_count)
-
-    for element_index in range(elements_count):  # TODO: prange?
-        element = elements[element_index]
-        element_points = nodes[element]
-
-        # TODO: Get rid of repetition (?)
-        for i in range(element_size):
-            i_dPhX, i_dPhY, element_volume = get_integral_parts(element_points, i)
-            # TODO: Avoid repetition
-            element_initial_volume[element_index] = element_volume
-
-            for j in range(element_size):
-                j_dPhX, j_dPhY, _ = get_integral_parts(element_points, j)
-
-                volume = (i != j) * (int_Ph / CONNECTED_EDGES_COUNT)
-                # divide by edge count - info about each triangle is "sent" to node via all connected edges
-                # (in 2D: 2, in 3D: 3) and summed (by dot product with matrix)
-                u = (1 + (i == j)) / 12.0
-                # in 3D: divide by 10 or 20, in 2D: divide by 6 or 12
-
-                v1 = int_Ph * j_dPhX
-                v2 = int_Ph * j_dPhY
-
-                w11 = i_dPhX * j_dPhX
-                w12 = i_dPhX * j_dPhY
-                w21 = i_dPhY * j_dPhX
-                w22 = i_dPhY * j_dPhY
-
-                edges_features_matrix[:, element[i], element[j]
-                ] += element_volume * np.array([volume, u, v1, v2, w11, w12, w21, w22])
-
-    return edges_features_matrix, element_initial_volume
-
-
 @njit
-def get_integral_parts(element_nodes, element_index):
-    x_i = element_nodes[element_index % 3]
-    x_j1 = element_nodes[(element_index + 1) % 3]
-    x_j2 = element_nodes[(element_index + 2) % 3]
-
-    dm = denominator(x_i, x_j1, x_j2)
-    triangle_area = np.abs(dm) / 2.0  # = np.abs(dm) / 2.0 = shoelace_area
-
-    y_sub = x_j2[1] - x_j1[1]
-    x_sub = x_j1[0] - x_j2[0]
-
-    dPhX = y_sub / dm
-    dPhY = x_sub / dm
-
-    return dPhX, dPhY, triangle_area
-
-
-@njit
-def shoelace_area(points):
-    x = points[:, 0].copy()
-    y = points[:, 1].copy()
-    area = 0.5 * np.abs(np.dot(x, np.roll(y, 1)) - np.dot(y, np.roll(x, 1)))
-    return area
-
-
-@njit
-def denominator(x_i, x_j1, x_j2):
-    return (
-        x_i[1] * x_j1[0]
-        + x_j1[1] * x_j2[0]
-        + x_i[0] * x_j2[1]
-        - x_i[1] * x_j2[0]
-        - x_j2[1] * x_j1[0]
-        - x_i[0] * x_j1[1]
-    )
-
-
-# @njit
-def get_edges_features_list(edges_number, edges_features_matrix):
+def get_edges_features_list_numba(edges_number, edges_features_matrix):
     nodes_count = len(edges_features_matrix[0])
     edges_features = np.zeros((edges_number + nodes_count, 8))  # , dtype=numba.double)
     e = 0
@@ -104,86 +22,6 @@ def get_edges_features_list(edges_number, edges_features_matrix):
                 e += 1
     return edges_features
 
-
-@njit
-def calculate_constitutive_matrices(W11, W12, W21, W22, MU, LA):
-    B11 = (2 * MU + LA) * W11 + MU * W22
-    B22 = MU * W11 + (2 * MU + LA) * W22
-    B12 = MU * W21 + LA * W12
-    B21 = LA * W21 + MU * W12
-    return B11, B12, B21, B22
-
-
-# @njit
-def get_matrices(
-    edges_features_matrix, MU, LA, TH, ZE, density, time_step, independent_nodes_count
-):
-    ind = slice(0, independent_nodes_count)
-    
-    VOL = edges_features_matrix[0]
-
-    U = edges_features_matrix[1][ind, ind]
-
-    V1 = edges_features_matrix[2][ind, ind]
-    V2 = edges_features_matrix[3][ind, ind]
-
-    W11 = edges_features_matrix[4]
-    W12 = edges_features_matrix[5]
-    W21 = edges_features_matrix[6]
-    W22 = edges_features_matrix[7]
-
-    A11, A12, A21, A22 = calculate_constitutive_matrices(W11, W12, W21, W22, TH, ZE)
-    B11, B12, B21, B22 = calculate_constitutive_matrices(W11, W12, W21, W22, MU, LA)
-
-    # A = np.vstack((np.hstack((c, s)), np.hstack((-s, c))))
-    # result = A @ matrix @ A.T
-
-
-    A = np.vstack(
-        (
-            np.hstack((A11[ind, ind], A12[ind, ind])),
-            np.hstack((A21[ind, ind], A22[ind, ind])),
-        )
-    )
-    B = np.vstack(
-        (
-            np.hstack((B11[ind, ind], B12[ind, ind])),
-            np.hstack((B21[ind, ind], B22[ind, ind])),
-        )
-    )
-
-
-    Z = np.zeros_like(U)
-    ACC = density * np.vstack((np.hstack((U, Z)), np.hstack((Z, U))))
-
-    A_plus_B_times_ts = A + B * time_step
-    C = ACC + A_plus_B_times_ts * time_step
-
-    """
-    k11 = 0.5
-    k12 = k21 = 0.5
-    k22 = 0.5
-    """
-
-    c11 = 1.5
-    c12 = c21 = 1.5
-    c22 = 1.5
-
-    c11 = c22 = 0.5
-    c12 = c21 = 0
-
-    C2X = c11 * V1 + c21 * V2
-    C2Y = c12 * V1 + c22 * V2
-
-    """
-    # T = (1.0 / TIMESTEP) * k11 * W11 + k12 * W12 + k21 * W21 + k22 * W22
-    """
-
-    k11 = k22 = 0.1
-    k12 = k21 = 0
-    K = k11 * W11 + k12 * W12 + k21 * W21 + k22 * W22
-
-    return C, B, VOL, A_plus_B_times_ts, A, ACC, K, C2X, C2Y
 
 
 class SettingMatrices(SettingMesh):
@@ -242,7 +80,7 @@ class SettingMatrices(SettingMesh):
         # edges_features = get_edges_features_list(
         #    self.edges_number, edges_features_matrix
         # )
-
+        slice_ind = slice(0, self.independent_nodes_count)
         (
             self.C,
             self.B,
@@ -261,7 +99,7 @@ class SettingMatrices(SettingMesh):
             self.ze,
             self.density,
             self.time_step,
-            self.independent_nodes_count,
+            slice_ind,
         )
 
         if self.with_C:
